libft

+++ client.c
@@ -3,45 +3,49 @@
 /*                                                        :::      ::::::::   */
 /*   client.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
+/*   By: yonshin <yonshin@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 09:07:44 by yonshin           #+#    #+#             */
-/*   Updated: 2022/11/16 20:40:01 by yonshin          ###   ########.fr       */
+/*   Updated: 2022/11/17 02:00:47 by yonshin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <signal.h>
-#include <stdio.h>
-#include <unistd.h>
+#include "minitalk.h"
 
-void	signal_char(int pid, unsigned char c)
+char	*g_str;
+void	signal_char(int pid)
 {
-	int	bit = 8;
-	while (bit--) {
-		int signal = SIGUSR1 + (c & 128) / 128;
-		printf("%d %x\n", signal, c);
+	static int	bit;
+	int			signal;
+
+	signal = SIGUSR1 + ((*g_str & (0b100000000 >> ++bit)) > 0);
+	if (bit != 8)
+	{
 		kill(pid, signal);
-		c = c << 1;
-		usleep(1000);
+		return ;
 	}
-}
-
-int	atoi(char *str)
-{
-	int	i = 0;
-	while (*str)
-		i = i * 10 + *str++ - '0';
-	return i;
+	if (*g_str == 0)
+	{
+		kill(pid, signal);
+		exit(0);
+	}
+	g_str += 1;
+	bit = 0;
+	kill(pid, signal);
 }
 
 int	main(int argc, char *argv[])
 {
+	int		pid;
+
 	if (argc < 3)
 		return (0);
-	int		pid = atoi(argv[1]);
-	char	*str = argv[2];
-
-	while (*str)
-		signal_char(pid, *str++);
+	pid = ft_atoi(argv[1]);
+	if (pid < 1)
+		return (1);
+	g_str = argv[2];
+	signal(SIGUSR1, signal_char);
+	kill(pid, SIGUSR1);
+	while (1) ;
 	return (0);
 }
